Hook with window.innerWidth
