+++ src/hooks/index.js
@@ -1,0 +1,1 @@
+export { useIsMobile } from "./use-is-mobile";
