+++ src/hooks/use-is-mobile.jsx
@@ -1,0 +1,19 @@
+import React, { useState, useEffect } from "react";
+
+const useIsMobile = () => {
+  const [width, setWidth] = useState(window.innerWidth);
+
+  useEffect(() => {
+    const handleWindowSizeChange = () => {
+      setWidth(window.innerWidth);
+    }
+    window.addEventListener('resize', handleWindowSizeChange);
+    return () => {
+      window.removeEventListener('resize', handleWindowSizeChange);
+    }
+  }, []);
+
+  return (width <= 768);
+};
+
+export { useIsMobile };
