+++ src/layouts/layout.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 
 import Helmet from "react-helmet";
 import { withPrefix, Link} from "gatsby";
@@ -14,6 +14,8 @@ import { StyledContainerHeader, StyledContainerNavBarXL } from "../elements";
 import "../index.scss";
 import { HomeBanner } from "../components/home-banner";
 import { AppContext, initialState, BannerType } from "../config/context.js";
+
+import { useIsMobile } from "../hooks";
 
 import * as styles from "./layout.module.scss";
 
@@ -59,21 +61,8 @@ export const StyledFooterText = styled.label`
 
 function Layout({ children }) {  
   const [contextState, setContextState] = useState(initialState);
-  const [width, setWidth] = useState(window.innerWidth);
   const [ handleMenu, setHandleMenu ] = useState(false);
-
-  const handleWindowSizeChange = () => {
-    setWidth(window.innerWidth);
-  }
-
-  useEffect(() => {
-    window.addEventListener('resize', handleWindowSizeChange);
-    return () => {
-      window.removeEventListener('resize', handleWindowSizeChange);
-    }
-  }, []);
-
-  const isMobile = width <= 768;
+  const isMobile = useIsMobile();
 
   const handleMenuNavShows = () => {
     if (isMobile === false) {
